Cambios en listas

+++ Intent.AnalisisDocumentos/BL/SearchFile.cs
@@ -7,7 +7,7 @@ namespace Intent.AnalisisDocumentos.BL
 {
     public class SearchFile
     {
-        private String[] files;
+        public List<String> files;
         private string extension;
 
         public List<string> Exitosos;
@@ -18,31 +18,40 @@ namespace Intent.AnalisisDocumentos.BL
             this.extension = extension;
             Exitosos = new List<string>();
             Fallidos = new List<string>();
-            files = Directory.GetFiles(searchPath, string.Format("{0}{1}", "*", extension), SearchOption.AllDirectories);
+            files = Directory.GetFiles(searchPath, string.Format("{0}{1}", "*", extension), SearchOption.AllDirectories).ToList();
         }
 
         public void SearchAccessibleFiles(string fileName)
         {
-            string file = files.FirstOrDefault(item => item.Contains(fileName));
-            FileStream fileStream = null;
-            if (!string.IsNullOrEmpty(file))
+            List<string> filess = (from item in files.AsParallel()
+                                 where item.Contains(fileName)
+                                 select item).ToList();
+            //string file = files.FirstOrDefault(item => item.Contains(fileName));
+            if (filess != null && filess.Count > 0)
             {
-                try
+                string file = files[0];
+                FileStream fileStream = null;
+                if (!string.IsNullOrEmpty(file))
                 {
-                    File.SetAttributes(file, FileAttributes.Normal);
-                    File.Delete(file);
-                    fileStream = File.Create(file);
-                    fileStream.Close();
-                    File.SetAttributes(file, FileAttributes.ReadOnly);
-                    Exitosos.Add(fileName);
+                    try
+                    {
+                        File.SetAttributes(file, FileAttributes.Normal);
+                        File.Delete(file);
+                        fileStream = File.Create(file);
+                        fileStream.Close();
+                        File.SetAttributes(file, FileAttributes.ReadOnly);
+                        Exitosos.Add(fileName);
+                    }
+                    catch (Exception)
+                    {
+                        Fallidos.Add(fileName);
+                    }
                 }
-                catch (Exception)
-                {
-                    Fallidos.Add(fileName);
-                }
+                //else
+                //    Fallidos.Add(fileName); 
             }
-            else
-                Fallidos.Add(fileName);
+            //else
+            //    Fallidos.Add(fileName); 
         }
 
         public void Log(List<string> exitosos, List<string> fallidos)
